convert svg and spec files
