+++ src/components/games/chess-bishops/helpers.spec.ts
