+++ src/components/games/chess-ducks/rubber-duck-svg.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 export const DuckSvg = () => {
   return (
     <svg style={{ display: 'none' }}>
