+++ src/components/games/chess-rook/bot-strategy.spec.ts
