+++ src/components/games/chess-rook/helpers.spec.ts
